+++ Example/AndesUI/Home/Presenters/HomePresenter.swift
@@ -22,7 +22,8 @@ protocol HomePresenter: NSObject {
     func presentCard()
     func presentThumbnail()
     func presentTextFieldsCode()
-    func presentListView()
+    func presentBottomSheet()
+    func presentList()
 }
 
 class HomeViewPresenter: NSObject {
@@ -37,6 +38,7 @@ class HomeViewPresenter: NSObject {
 }
 
 extension HomeViewPresenter: HomePresenter {
+
     func presentCoachmark() {
         router.routeToCoachmark()
     }
@@ -82,11 +84,15 @@ extension HomeViewPresenter: HomePresenter {
         router.routeToThumbnail()
     }
 
-    func presentListView() {
+    func presentList() {
         router.routeToList()
     }
 
     func presentTextFieldsCode() {
         router.routeToTextFieldsCode()
+    }
+
+    func presentBottomSheet() {
+        router.routeToBottomSheet()
     }
 }